create unitTests directory and move unit tests (don't depend on SwingSet)
