+++ test/unitTests/more/pixels/types/test-area.js
