+++ test/unitTests/more/pixels/types/test-pixel.js
